+++ Source/SpriteSort2DEditor/Private/SpriteSortComponentCustomization.h
@@ -6,11 +6,6 @@
 #include "UObject/WeakObjectPtrTemplates.h"
 
 class USpriteSortComponent;
-class UBlueprint;
-class UPrimitiveComponent;
-class USceneComponent;
-class USCS_Node;
-class USimpleConstructionScript;
 
 class FSpriteSortComponentCustomization final : public IDetailCustomization
 {
@@ -22,16 +17,6 @@ public:
 private:
 	TArray<TWeakObjectPtr<USpriteSortComponent>> Components;
 
-	FReply UpdateSortNow();
-	FReply AutoFindVisualRoot();
-	FReply CreateSortOrigin();
 	FText GetWarningText() const;
 	EVisibility GetWarningVisibility() const;
-
-	bool AutoFindOnBlueprintTemplate(USpriteSortComponent* Component) const;
-	bool CreateSortOriginOnBlueprintTemplate(USpriteSortComponent* Component) const;
-	UBlueprint* GetBlueprintForTemplate(const USpriteSortComponent* Component) const;
-	USCS_Node* FindNodeForTemplate(const USpriteSortComponent* Component) const;
-	USceneComponent* FindBestBlueprintVisualRoot(USimpleConstructionScript* SCS) const;
-	UPrimitiveComponent* FindBestBlueprintTargetPrimitive(USimpleConstructionScript* SCS, USceneComponent* VisualRoot) const;
 };